Catch referenced errors in IFERROR
Formula references evaluate to ref nodes first, so IFERROR and IFNA were missing error sentinels already stored in referenced cells. Reuse the ERROR.TYPE error-value helper path so referenced formula errors are detected consistently.
